+++ NEWS.md
@@ -1,4 +1,4 @@
-# TreeTools 1.7.2.9000 #
+# TreeTools 1.7.3 #
 
 - `ReadTntTree()` supports multi-line trees.